Creating View class
